# Calculate insight score once per time step