--- a/Engine/Alphas/DefaultAlphaHandler.cs
+++ b/Engine/Alphas/DefaultAlphaHandler.cs
@@ -38,8 +38,8 @@ namespace QuantConnect.Lean.Engine.Alphas
     /// </summary>
     public class DefaultAlphaHandler : IAlphaHandler
     {
-        private DateTime _lastSecurityValuesSnapshotTime;
-
+        private DateTime _lastStepTime;
+        private List<Insight> _insights;
         private ChartingInsightManagerExtension _charting;
         private ISecurityValuesProvider _securityValuesProvider;
         private CancellationTokenSource _cancellationTokenSource;
@@ -102,6 +102,7 @@ namespace QuantConnect.Lean.Engine.Alphas
             MessagingHandler = messagingHandler;
 
             _fitnessScore = new FitnessScoreManager();
+            _insights = new List<Insight>();
             _securityValuesProvider = new AlgorithmSecurityValuesProvider(algorithm);
 
             InsightManager = CreateInsightManager();
@@ -116,7 +117,13 @@ namespace QuantConnect.Lean.Engine.Alphas
             InsightManager.AddExtension(_charting);
 
             // when insight is generated, take snapshot of securities and place in queue for insight manager to process on alpha thread
-            algorithm.InsightsGenerated += (algo, collection) => InsightManager.Step(collection.DateTimeUtc, CreateSecurityValuesSnapshot(), collection);
+            algorithm.InsightsGenerated += (algo, collection) =>
+            {
+                lock (_insights)
+                {
+                    _insights.AddRange(collection.Insights);
+                }
+            };
         }
 
         /// <summary>
@@ -138,9 +145,14 @@ namespace QuantConnect.Lean.Engine.Alphas
         public virtual void ProcessSynchronousEvents()
         {
             // check the last snap shot time, we may have already produced a snapshot via OnInsightssGenerated
-            if (_lastSecurityValuesSnapshotTime != Algorithm.UtcTime)
+            if (_lastStepTime != Algorithm.UtcTime)
             {
-                InsightManager.Step(Algorithm.UtcTime, CreateSecurityValuesSnapshot(), null);
+                _lastStepTime = Algorithm.UtcTime;
+                lock (_insights)
+                {
+                    InsightManager.Step(_lastStepTime, _securityValuesProvider.GetAllValues(), new GeneratedInsightsCollection(_lastStepTime, _insights, clone: false));
+                    _insights.Clear();
+                }
             }
 
             if (_lastFitnessScoreCalculation.Date != Algorithm.UtcTime.Date)
@@ -256,12 +268,6 @@ namespace QuantConnect.Lean.Engine.Alphas
         protected virtual AlphaResultPacketSender CreateAlphaResultPacketSender()
         {
             return new AlphaResultPacketSender(Job, MessagingHandler, TimeSpan.FromSeconds(1), 50);
-        }
-
-        private ReadOnlySecurityValuesCollection CreateSecurityValuesSnapshot()
-        {
-            _lastSecurityValuesSnapshotTime = Algorithm.UtcTime;
-            return _securityValuesProvider.GetAllValues();
         }
 
         /// <summary>