<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Calculate insight scores once per time step.
    - Insights scores will be calculated at the end of each time step so that insights closed, because of new order based insights, get updated in the same time step they are closed.
- Regression tests statistics change slightly: the rolling average (an EMA) changed because the order of the closed insights is different, this is because insights are stored in a hash set which does not guarantee a specify order (each insight score is still the same as `master`).

`Performance tests StatefulCoarseUniverseSelectionBenchmark 2017 - 2018 | 250 securities` 
`master`
38.83 seconds at 50k data points per second. Processing total of 1,925,402 data points.
38.49 seconds at 50k data points per second. Processing total of 1,925,402 data points.
38.44 seconds at 50k data points per second. Processing total of 1,925,402 data points.
`PR: ~15% improvement`
34.12 seconds at 56k data points per second. Processing total of 1,925,402 data points.
33.68 seconds at 57k data points per second. Processing total of 1,925,402 data points.
33.45 seconds at 58k data points per second. Processing total of 1,925,402 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3298
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- Performance
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->